Fix bug where the missing date caused an entry display nonsense.
Resolves (#1)
Added a `parseLanguage` filter function which can be used if an item attribute is not a string but a object of the form:
``` js
{'en': 'some english text', 'de': 'irgendein deutscher Text'}
```
The `parseLanguage` function will check the current browser language and try to parse the correct output.

Changed some ES4 to ES6 variable declarations.
Add documentation on how to test the widget in a locally run cockpit.
Update CHANGELOG.md
Added `.idea` folder to .gitignore